[1] refactor code. [2] add exception handling to handle overload of queue requests/response from twitter streaming api

+++ TwitterAPIStreaming.py
@@ -3,7 +3,8 @@ from MyStreamListener import *
 from socket import *
 import os
 import sys
-
+from http.client import IncompleteRead # Python 3
+from urllib3.exceptions import ProtocolError
 
 class TwitterAPIStreaming:
     streaming_duration = 0.0
@@ -13,31 +14,49 @@ class TwitterAPIStreaming:
 
     # Map from tweets in file to SMS objects
 
-    def stream_tweets(self, output_file, duration, filter_language, serialize_option):
+    def stream_tweets(self, output_file, duration, filter_language, serialize_option, full_spritzer):
+        
         start_time = time.time()
         self.streaming_duration = duration
 
-        print("\nST: Streaming Tweets")
-        print("streaming duration : %s seconds" % self.streaming_duration)
+        print()
+        if (full_spritzer):
+            print("ST: Streaming full twitter spritzer (sample)")
+        else:
+            print("ST: Streaming duration : %s seconds" % self.streaming_duration)
+        print("ST: Streaming Tweets in progress ... ")
 
-        try:
-            # tweets_public
-            # [1] Create Listener. [2] Initialize data
-            stream_listener = MyStreamListener(tweet_messages_pb=output_file, filter_lang=filter_language,
-                                               duration=self.streaming_duration, start_time=start_time,
-                                               option=serialize_option)
+        continue_st = True
+        while continue_st :
+            try:
+                # tweets_public
+                # [1] Create Listener. [2] Initialize data
+                stream_listener = MyStreamListener(tweet_messages_pb=output_file, filter_lang=filter_language,
+                                                   duration=self.streaming_duration, start_time=start_time,
+                                                   option=serialize_option, full_spritzer=full_spritzer)
 
-            # [3] Start the stream with the specified listener class
-            my_stream = tweepy.Stream(auth=self.apiHandle.auth, listener=stream_listener)
+                # [3] Start the stream with the specified listener class
+                my_stream = tweepy.Stream(auth=self.apiHandle.auth, listener=stream_listener)
 
-            # [4] Tell the stream listener to retrieve (filter) only the twitter spritzer sample
-            MyStreamListener(my_stream.sample(), filter_lang=filter_language,
-                             duration=self.streaming_duration, start_time=start_time, option=serialize_option)
-            return "done"
+                # [4] Tell the stream listener to retrieve (filter) only the twitter spritzer sample
+                MyStreamListener(my_stream.sample(), filter_lang=filter_language,
+                                 duration=self.streaming_duration, start_time=start_time, 
+                                 option=serialize_option, full_spritzer=full_spritzer)
 
-        except tweepy.TweepError as te_error:
-            my_stream.disconnect()
-            print("Get_tweets error: ", te_error.__str__())
+                return "done"
+            except ProtocolError:
+                print ("TAPIS: ProtocolError ... pass")
+                continue
+
+            except IncompleteRead:
+                print ("TAPIS: IncompleteRead ... pass")
+                continue
+
+            except tweepy.TweepError as te_error:
+                my_stream.disconnect()
+                print("Get_tweets error: ", te_error.__str__())
+            continue_st = False
+            # break
 
     """send message object"""
     @staticmethod
@@ -53,43 +72,44 @@ class TwitterAPIStreaming:
             # connect to server
             client_socket.connect((d_server, d_port))
 
-            # send data to server
-            # Protocol buffer
-            if serialize_option == "1":
+            # Protocol buffer 
+            if serialize_option == 1:
                 pb_file = open(pb_serialized_file, "rb")
-                f_data = pb_file.read()
+                file_data = pb_file.read()
                 pb_file.close()
 
             # Json
             elif serialize_option == 2:
-                pb_file = open(pb_serialized_file, "r")
-                f_data = pb_file.read()
-                pb_file.close()
+                json_file = open(pb_serialized_file, "r")
+                file_data = json_file.read()
+                json_file.close()
 
-            client_socket.send(f_data)
+            # send data to server
+            client_socket.send(file_data)
             client_socket.close()
-            return "success"
+            return "file sent!"
 
         except Exception as s_error:
             print("Error send: " + s_error.__str__())
-            return "failed"
+            return "file not sent!"
 
-    def reformat_Json(self, filename):
-        # if os.stat(filename).st_size == 0:
-        #     print("file not found")
+    def reformat_json(self, filename):
+        """Correct json data format. Insert all items in a list"""
+        if os.stat(filename).st_size == 0:
+            print("file not found")
 
-        with open(filename, "r+") as rw_file:
-            """"""
-            json_list = []
-            for item in rw_file.read().split("\n"):
-                if len(item) > 0:
-                    json_list.append(json.loads(item))
+        else:
+            with open(filename, "r+") as rw_file:
+                """"""
+                json_list = []
+                for item in rw_file.read().split("\n"):
 
-            # print("List -> ", json_list.__dict__)
+                    if len(item) > 0:
+                        json_list.append(json.loads(item))
 
-            rw_file.seek(0)
-            json.dump(json_list, rw_file)
-            rw_file.truncate()
+                rw_file.seek(0)
+                json.dump(json_list, rw_file, indent=4)
+                rw_file.truncate()
 
     @staticmethod
     def deserialize_messages(binary_file, ser_option):
@@ -97,6 +117,7 @@ class TwitterAPIStreaming:
         This method retrieves the tweets data from the protocol buffer
 
         :param binary_file:
+        :param ser_option
         :return: sms_list
         """
 
@@ -117,7 +138,6 @@ class TwitterAPIStreaming:
 
                 try:
                     for anSMS in json_obj:
-                        print(anSMS)
                         SMSobj = SMS(anSMS["tweet_id"], anSMS["date"], anSMS["tweet_text"], anSMS["user_id"],
                                      anSMS["user_name"], anSMS["user_screen_name"], anSMS["source"],
                                      anSMS["user_location"], anSMS["text_lan"], anSMS["message_id"])
@@ -139,76 +159,81 @@ class TwitterAPIStreaming:
 
 
 def main():
-    """ ================ INPUT VARIABLES ================ """
     """ Define application key, key-secret and access-token """
     consumer_key = "g104Mx068t8JqSOfCCGPJ5xhx"
     consumer_key_secret = "suuzYhQCAQQJAT2K4LqAOVWeB9ZtKPp01jFA8KDPUGTdoYgc5l"
     access_token = "937943519985094657-W0MJHdYgUoJ4eHqleO4HpUedYjjprZI"
     access_token_secret = "ZGiEnXPCJbrNFmgBImKyUG55ImNXanVllpwQ0Nnt1K1Jp"
 
-    """Get this input from user"""
-    # command args or input
     # if len(sys.argv) != n: print "usage: python 'TwitterAPIStreaming.py 'pb filename''"
     # file = open(sys.argv[1], "rb")
-
     filename = "media/twitter_spritzer"
-    streaming_duration = float(10 * 1)  # stream tweets for n minutes
     filter_languages = ["en"]
-    dest_server = "41.193.221.138"
-    dest_port = 80
+    full_spritzer_i = False
+    option = 0
+    streaming_duration = float(0) # Enter 0 for a full spritzer
+
+    """ ================ USER INPUT  ================ """
+    print()
+    print("Please enter following inputs (blank for default test values)... ")
+    dest_server = input (" * File Destination Server-address : ") 
+    dest_port = input (" * File Destination Port-number : ") 
+    lang = input (" * Tweets Language Code [en, af,] seperated by comma (leave blank for 'en-english') : ") 
 
     try:
         option = \
-            input("\nChoose a serialising option (enter a number): \n 1. protocol buffer \n 2. json \n\nOption:")
+            input("\nChoose a serialising option (enter a number): \n 1. Protocol buffer \n 2. Json \n\nOption : ")
         option = int(option)
-
     except Exception:
-        print("Error : Please enter a number")
+        print("Error : Option should be a number: 1 or 2")
+
+    for lang_i in lang.split(","):
+        filter_languages.append(lang_i)
+
+    if streaming_duration == float(0):
+        full_spritzer_i = True
+   
+    if dest_server == "":
+        dest_server = "localhost"
+    if dest_port == 0 or dest_port=="":
+        dest_port = 80
 
     if option == 1:
-        filename += ".bin"
-        print("... Serialising set to Protocol buffer")
+        filename += ".bin" 
+        print("... Data Serialising set to Protocol buffer.")
     elif option == 2:
         filename += ".json"
-        print("... Serialising set to Json")
+        print("... Data Serialising set to Json.")
     else:
-        print("Option not found. Please enter option 1 or 2")
+        print("Option not found. Please enter option 1 or 2.")
         exit()
 
-    print("\nFilename : ", filename)
-    print("File destination : [ Server= %s, Port= %s ]" % (dest_server, dest_port))
-
-    # delete the file if it exists
+    # Start clean. Delete the file if it already exists
     try:
         os.remove(filename)
-        print("\n... File already exists. Deleted!")
+        # print("\n... File already exists. Deleted!")
     except OSError:
+        print("\n[Remove]: File not found")
         pass
-
-    if len(filename) == 0:
-        print("ST error: Tweets output file missing!")
-        exit()
-    if streaming_duration < 10:
-        print("ST error: Duration must at least be >= 10")
-        exit()
 
     """ Authenticate application with the application keys """
     apiHandle = Authenticate().authenticate_app(consumer_key, consumer_key_secret, access_token, access_token_secret)
     query_stream = TwitterAPIStreaming(apiHandle)
 
-    # query_stream.flushFile(filename)
-
     """ Get tweets from Streaming API and Map tweets/SMSes to protocol buffers """
-    query_stream.stream_tweets(output_file=filename, duration=streaming_duration, filter_language=filter_languages,
-                               serialize_option=option)
+    query_stream.stream_tweets(output_file=filename, duration=float(streaming_duration), filter_language=filter_languages,
+                                   serialize_option=option, full_spritzer=full_spritzer_i)
+    
+    print("\nAPI Streaming complete :)")
 
     if option == 2:
-        query_stream.reformat_Json(filename)
+        query_stream.reformat_json(filename)
+
+    print("\nData serialized to : ", filename)
+    print("File destination : [ Server = %s, Port = %s ]" % (dest_server, dest_port))
 
     # """ Deserialize messages """
-    query_stream.deserialize_messages(binary_file=filename, ser_option=option)
-
-    # SMS storage & config
+    # query_stream.deserialize_messages(binary_file=filename, ser_option=option)
 
     """ SMS transfer """
     # query_stream.send_serialized_file(filename, d_server=dest_server, d_port=dest_port, option=option)
